Chore(registerTalks): movendo para pasta separada com json, e factorie de admin
