--- src/bot/commands/compet/newTalks/new-talks-embed-error.json
@@ -1,18 +1,0 @@
-{
-    "color": 16082277,
-    "title": "Não foi possível utilizar este comando!",
-    "description": "Você não possui autorização necessária.",
-    "thumbnail": "https://www.pngfind.com/pngs/m/0-1420_red-cross-mark-clipart-green-checkmark-red-x.png",
-    "fields": [
-        {
-            "name": "Código do erro",
-            "value": "401",
-            "inline": false
-        },
-        {
-            "name": "Mensagem do erro",
-            "value": "Você precisa ter permissão de administrador para executar esse comando",
-            "inline": false
-        }
-    ]
-}
